Cleanup lib/cpalgo/geometry/***

+++ lib/cpalgo/geometry/vector2.hpp
@@ -257,8 +257,8 @@ vector2 intersection_ss(
 // Verified https://onlinejudge.u-aizu.ac.jp/courses/library/4/CGL/3/CGL_3_A
 double polygon_area(const std::vector<vector2>& G) {
     double ans = 0;
-    for (int i = 0; i < G.size(); ++i) {
-        int j = (i + 1) % G.size();
+    for (size_t i = 0; i < G.size(); ++i) {
+        size_t j = (i + 1) % G.size();
         ans += cross(G[i], G[j]);
     }
     return ans / 2;
@@ -268,9 +268,9 @@ double polygon_area(const std::vector<vector2>& G) {
 // The coordinates of points are given in the order of counter-clockwise visit of them.
 // Verified https://onlinejudge.u-aizu.ac.jp/courses/library/4/CGL/3/CGL_3_B
 bool is_convex_polygon(const std::vector<vector2>& G) {
-    for (int i = 0; i < G.size(); ++i) {
-        int p = (i-1+G.size()) % G.size();
-        int n = (i+1) % G.size();
+    for (size_t i = 0; i < G.size(); ++i) {
+        size_t p = (i-1+G.size()) % G.size();
+        size_t n = (i+1) % G.size();
         int c = ccw(G[p], G[i], G[n]);
         if (c == -1) return false;
     }
@@ -285,8 +285,8 @@ bool is_convex_polygon(const std::vector<vector2>& G) {
 int polgygon_contains(const std::vector<vector2> G, const vector2& p) {
     enum { ON = 0, IN = 1, OUT = -1 };
     bool in = false;
-    for (int i = 0; i < G.size(); ++i) {
-        int j = (i+1) % G.size();
+    for (size_t i = 0; i < G.size(); ++i) {
+        size_t j = (i+1) % G.size();
         auto a = G[i] - p, b = G[j] - p;
         if (EQ(cross(a,b),0) && LQ(dot(a,b), 0)) return ON;
         if (a.y > b.y) {
